Moved code FinalEngineGL to finalEngine.cpp/.h
