+++ source/blender/render/hydra/CMakeLists.txt
@@ -49,8 +49,6 @@ set(SRC
   engine.cpp
   finalEngine.h
   finalEngine.cpp
-  finalEngineGL.h
-  finalEngineGL.cpp
   viewportEngine.h
   viewportEngine.cpp
